requirements.txt
black==22.10.0
click==8.1.3
colorama==0.4.6
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
prompt-toolkit==1.0.14
Pygments==2.13.0
PyInquirer==1.0.3
python-dotenv==0.21.0
regex==2022.10.31
six==1.16.0
tomli==2.0.1
wcwidth==0.2.5
